Move authflowclient to directly under lib
To signify that it does not depend on pkg/lib/authenticationflow
